Layout, usability, style, etc..

+++ iloire Facturacion/Controllers/InvoiceDetailsController.cs
@@ -20,6 +20,8 @@ namespace iloire_Facturacion.Controllers
             var invoicedetails = db.InvoiceDetails.Include(i => i.Invoice).Where(i=>i.InvoiceID==id);
             return PartialView("Index", invoicedetails.ToList());
         }
+
+    
         /*END CUSTOM*/
 
         //
@@ -46,27 +48,35 @@ namespace iloire_Facturacion.Controllers
         public ActionResult Create(int? id)
         {
             ViewBag.InvoiceID = new SelectList(db.Invoices, "InvoiceID", "Notes");
+            Invoice invoice = null;
+            InvoiceDetails i = null;
 
             if (id.HasValue) {
-                var invoice = (from i in db.Invoices
-                               where i.InvoiceID == id
-                               select i).FirstOrDefault();
-                
-                if (invoice != null) {
-                    InvoiceDetails i = new InvoiceDetails();
+                invoice = (from ii in db.Invoices
+                               where ii.InvoiceID == id
+                               select ii).FirstOrDefault();
+
+                if (invoice != null)
+                {
+                    i = new InvoiceDetails();
                     i.InvoiceID = id.Value;
                     i.Invoice = invoice;
+                    i.Qty = 1;
+                    i.VAT = 18;
 
                     i.TimeStamp = DateTime.Now;
 
                     ViewBag.InvoiceID = new SelectList(db.Invoices, "InvoiceID", "Notes", id.Value);
-                    return View(i);
                 }
-                else
-                    return View();
             }
-            else
-                return View();
+
+            if (Request.IsAjaxRequest())
+            {
+                return PartialView("Create", i);
+            }
+            else {
+                return View("Create", i);
+            }
         } 
 
         //
@@ -79,21 +89,22 @@ namespace iloire_Facturacion.Controllers
             {
                 db.InvoiceDetails.Add(invoicedetails);
                 db.SaveChanges();
-                return RedirectToAction("Details", "Invoice", new { id = invoicedetails.InvoiceID });
+                return PartialView("Index", db.InvoiceDetails.Where(i => i.InvoiceID == invoicedetails.InvoiceID));
             }
 
             ViewBag.InvoiceID = new SelectList(db.Invoices, "InvoiceID", "Notes", invoicedetails.InvoiceID);
-            return View(invoicedetails);
+            this.Response.StatusCode = 400;
+            return PartialView("Create", invoicedetails);
         }
         
         //
         // GET: /InvoiceDetails/Edit/5
- 
+
         public ActionResult Edit(int id)
         {
             InvoiceDetails invoicedetails = db.InvoiceDetails.Find(id);
             ViewBag.InvoiceID = new SelectList(db.Invoices, "InvoiceID", "Notes", invoicedetails.InvoiceID);
-            return View(invoicedetails);
+            return PartialView(invoicedetails);
         }
 
         //
@@ -106,10 +117,11 @@ namespace iloire_Facturacion.Controllers
             {
                 db.Entry(invoicedetails).State = EntityState.Modified;
                 db.SaveChanges();
-                return RedirectToAction("Details", "Invoice", new  { id = invoicedetails.InvoiceID });
-            }
+                return RedirectToAction("IndexByInvoice", "InvoiceDetails", new { id = invoicedetails.InvoiceID });
+            }       
             ViewBag.InvoiceID = new SelectList(db.Invoices, "InvoiceID", "Notes", invoicedetails.InvoiceID);
-            return View(invoicedetails);
+            this.Response.StatusCode = 400;
+            return PartialView("Edit", invoicedetails);
         }
 
         //
@@ -118,7 +130,7 @@ namespace iloire_Facturacion.Controllers
         public ActionResult Delete(int id)
         {
             InvoiceDetails invoicedetails = db.InvoiceDetails.Find(id);
-            return View(invoicedetails);
+            return PartialView(invoicedetails);
         }
 
         //
@@ -128,9 +140,16 @@ namespace iloire_Facturacion.Controllers
         public ActionResult DeleteConfirmed(int id)
         {            
             InvoiceDetails invoicedetails = db.InvoiceDetails.Find(id);
-            db.InvoiceDetails.Remove(invoicedetails);
-            db.SaveChanges();
-            return RedirectToAction("Details", "Invoice", new { id = invoicedetails.InvoiceID });
+            if (invoicedetails != null)
+            {
+                db.InvoiceDetails.Remove(invoicedetails);
+                db.SaveChanges();
+                return RedirectToAction("IndexByInvoice", "InvoiceDetails", new { id = invoicedetails.InvoiceID });
+            }
+            else {
+                this.Response.StatusCode = 400;
+                return Content("Record not found");
+            }
         }
 
         protected override void Dispose(bool disposing)
